+++ lib/rack/reducer/middleware.rb
@@ -1,10 +1,9 @@
 # frozen_string_literal: true
 
 require 'rack/request'
-require_relative 'reduction'
 
 module Rack
-  module Reducer
+  class Reducer
     # Mount Rack::Reducer as middleware
     # @example A microservice that filters artists
     #   ArtistService = Rack::Builder.new do
@@ -23,10 +22,10 @@ module Rack
       def initialize(app, options = {})
         @app = app
         @key = options[:key] || 'rack.reduction'
-        @reducer = Rack::Reducer.create(options[:dataset], *options[:filters])
+        @reducer = Rack::Reducer.new(options[:dataset], *options[:filters])
       end
 
-      # Call the next app in the middleware stack, with env[key] set
+      # Call the next app in the middleware stack, with `env[key]` set
       # to the ouput of a reduction
       def call(env)
         params = Rack::Request.new(env).params